SWA-generator
Here is SWA-generator that I came up with. It works by merging only certain layers that are similar to each other. Info on what layers are similar is written to a file (here called report82-83.txt). Here the user must decide what layers to merge, and write them into the program (this part is commented).  The new network will be written to file 83-82.txt

Here you can see that layers 64, 58, 56, 55, 52, 48 and 44 have already been selected. This was found to produce a network that defeated net 83 20-10 in a thirty game match. This match was played with a 10 ply opening book, engines played both sides with 0.5sec per move (~1000 playouts per move on my computer)

Meaning of the terms in the report:

No. = Layer number. There are 65 'layers'. This is because numbers after a newline character are treated as a new layer. In general, you should only merge layers larger than ~40. Merging layers as low as 11 would probably be disastrous. 

FR = Flip Ratio, this is the number of weights that have retained their sign divided by the number of weights that have changed their sign. High flip ratios are good, as it shows layers are similar. Preferably FR should be greater than 100.

%MD and %AD = % Multiplicative Difference and % Additive Difference. You can read exactly how they are calculated in the code, but in general, they measure the difference in corresponding weight sizes between the layers. %AD puts more emphasis on the larger weights. I find that %AD is a much better predictor of compatibility than %MD. In general, %AD < 3 is pretty good. 

Size = How many weights are in the layer.

The program suggests interesting layers to merge at the top but does include the others lower down in the program. Knowing what layers to merge is pretty difficult though, and there’s probably more to it than what I know. Others can maybe learn and create better indicators than I’ve given here. 

I hope you guys like SWA, as it can be an easy avenue to improve as it requires very little computational effort.